+++ src/exec/CssExecutor.ts
@@ -16,11 +16,11 @@ export class CssExecutor extends SimpleExecutor {
 
     // 注入 CSS 样式到文档头部
     injectStyle(id: string, code: string) {
-        let el = document.getElementById(`pluto-css-${id}`);
+        let el = activeDocument.getElementById(`pluto-css-${id}`);
         if (!el) {
-            el = document.createElement('style');
+            el = activeDocument.createElement('style');
             el.id = `pluto-css-${id}`;
-            document.head.appendChild(el);
+            activeDocument.head.appendChild(el);
         }
         el.textContent = code;
     }
@@ -34,6 +34,6 @@ export class CssExecutor extends SimpleExecutor {
     }
 
     async uninstall(context: BatchContext): Promise<void> {
-        document.querySelectorAll(`[id^="pluto-css-${context.id}-"]`).forEach(el => el.remove());
+        activeDocument.querySelectorAll(`[id^="pluto-css-${context.id}-"]`).forEach(el => el.remove());
     }
 }